check_akamai_bandwidth version bump...
Following a series of rather odd events our Akamai account expiration
went negative, yet it never stopped working. This change should allow
this, although it will end up straight to UNKNOWN if the account then
stop working (which is what the check was for in the first place).
I will properly detect and handle negative expiration if it turns
out to be only temporarily working.

I'm not sure if your Akamai rep will willfully enable an already
expired account but if you get it done this will avoid expiration
alerts.